chore: rename format url to embed
